pyproject.toml
[project]
name = "using-python-in-science"
version = "0.1.0"
description = ""
authors = [
    {name = "Jakub Wawrzonkowski",email = "j.wawrzon01@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "rich (>=13.9.4,<14.0.0)",
    "ascii-graph (>=1.5.1,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "numba (>=0.61.0,<0.62.0)",
    "requests (>=2.32.3,<3.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "selenium (>=4.28.1,<5.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)",
    "matplotlib (>=3.10.0,<4.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "ipython (>=8.31.0,<9.0.0)",
    "sympy (>=1.13.3,<2.0.0)",
    "bokeh (>=3.6.2,<4.0.0)"
]

[tool.poetry]
package-mode = false



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
